requirements.txt
flask
groq
numpy
scikit-learn
sentence-transformers
python-dotenv